# openapi/core-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/matching/invites": {
      "post": {
        "operationId": "InviteController_createInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateInviteDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/MatchInviteDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mời Voice/Soul Match trực tiếp tới 1 user — 409 nếu đã có lời mời đang chờ tới đúng người này",
        "tags": ["matching"]
      },
      "get": {
        "operationId": "InviteController_listReceived",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số item tối đa mỗi trang (1-100, mặc định 20)",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaque từ `meta.nextCursor` của trang trước",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MatchInvitesPageDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Danh sách lời mời ĐANG CHỜ phản hồi gửi tới chính mình",
        "tags": ["matching"]
      }
    },
    "/api/v1/matching/invites/{id}": {
      "get": {
        "operationId": "InviteController_getInvite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/MatchInviteDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Chi tiết 1 lời mời — chỉ inviter/invitee xem được",
        "tags": ["matching"]
      }
    },
    "/api/v1/matching/invites/{id}/accept": {
      "post": {
        "operationId": "InviteController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MatchInviteAcceptedDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Chấp nhận lời mời — tạo trực tiếp ticket/session, gọi confirmTicket(inviteeTicketId) tiếp theo như auto-match",
        "tags": ["matching"]
      }
    },
    "/api/v1/matching/invites/{id}/decline": {
      "post": {
        "operationId": "InviteController_decline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/MatchInviteDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Từ chối lời mời — chỉ invitee",
        "tags": ["matching"]
      }
    },
    "/api/v1/matching/invites/{id}/cancel": {
      "post": {
        "operationId": "InviteController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/MatchInviteDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Huỷ lời mời đã gửi — chỉ inviter, trước khi có phản hồi",
        "tags": ["matching"]
      }
    },
...
    "/api/v1/mood/presets": {
      "get": {
        "operationId": "MoodController_listPresets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MoodPresetDto" }
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Catalog mood preset đang bật",
        "tags": ["mood"]
      }
    },
    "/api/v1/mood/status/me": {
      "get": {
        "operationId": "MoodController_getMyStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MoodStatusResponseDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mood hiện tại của chính mình",
        "tags": ["mood"]
      }
    },
    "/api/v1/mood/status/{userId}": {
      "get": {
        "operationId": "MoodController_getStatus",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MoodStatusResponseDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mood công khai của 1 user — ẩn nếu có block 2 chiều (docs/services/mood-service.md § 3)",
        "tags": ["mood"]
      }
    },
    "/api/v1/mood/status": {
      "post": {
        "operationId": "MoodController_setStatus",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SetMoodDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MoodStatusResponseDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Set mood theo preset — auto-approve (W1 preset-only)",
        "tags": ["mood"]
      },
      "delete": {
        "operationId": "MoodController_clearStatus",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Tắt mood hiện tại",
        "tags": ["mood"]
      }
    }
  },
  "info": {
    "title": "Litmatch core-api",
    "description": "",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "PublicProfileDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "nickname": { "type": "string" },
          "gender": {
            "type": "string",
            "enum": ["unknown", "male", "female", "other"]
          },
          "avatarId": { "type": "string" },
          "interests": {
            "nullable": true,
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["id", "nickname", "gender", "avatarId", "interests"]
      },
...
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "inviteeUserId": {
            "type": "string",
            "description": "userId người được mời"
          },
          "matchType": {
            "type": "string",
            "enum": ["soul", "voice"],
            "example": "voice"
          }
        },
        "required": ["inviteeUserId", "matchType"]
      },
      "MatchInviteDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "inviterUserId": { "type": "string" },
          "inviteeUserId": { "type": "string" },
          "matchType": { "type": "string", "enum": ["soul", "voice"] },
          "status": {
            "type": "string",
            "enum": ["pending", "accepted", "declined", "expired", "cancelled"]
          },
          "expiresAt": { "format": "date-time", "type": "string" },
          "sessionId": { "type": "string", "nullable": true },
          "createdAt": { "format": "date-time", "type": "string" },
          "inviterProfile": {
            "description": "Profile công khai tối thiểu của người mời để người nhận quyết định có chủ đích.",
            "allOf": [{ "$ref": "#/components/schemas/PublicProfileDto" }]
          }
        },
        "required": [
          "id",
          "inviterUserId",
          "inviteeUserId",
          "matchType",
          "status",
          "expiresAt",
          "sessionId",
          "createdAt",
          "inviterProfile"
        ]
      },
      "MatchInvitesPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MatchInviteDto" }
          },
          "nextCursor": { "type": "string", "nullable": true }
        },
        "required": ["items", "nextCursor"]
      },
      "MatchInviteAcceptedDto": {
        "type": "object",
        "properties": {
          "invite": { "$ref": "#/components/schemas/MatchInviteDto" },
          "sessionId": {
            "type": "string",
            "description": "sessionId — cùng khái niệm session của auto-match"
          },
          "inviteeTicketId": {
            "type": "string",
            "description": "ticketId CỦA CHÍNH NGƯỜI ACCEPT — gọi confirmTicket(ticketId) như auto-match"
          }
        },
        "required": ["invite", "sessionId", "inviteeTicketId"]
      },
...
      "MoodPresetDto": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "label": { "type": "string" },
          "emoji": { "type": "string" }
        },
        "required": ["code", "label", "emoji"]
      },
      "CurrentMoodDto": {
        "type": "object",
        "properties": {
          "preset": { "$ref": "#/components/schemas/MoodPresetDto" },
          "setAt": { "format": "date-time", "type": "string" },
          "expiresAt": { "format": "date-time", "type": "string" }
        },
        "required": ["preset", "setAt", "expiresAt"]
      },
      "MoodStatusResponseDto": {
        "type": "object",
        "properties": {
          "mood": {
            "nullable": true,
            "type": "object",
            "allOf": [{ "$ref": "#/components/schemas/CurrentMoodDto" }]
          }
        }
      },
      "SetMoodDto": {
        "type": "object",
        "properties": {
          "presetCode": { "type": "string", "example": "happy" }
        },
        "required": ["presetCode"]
      }
    }
  }
}